hotelTuG/requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.3
django-mysql==4.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.13.0
mysqlclient==2.1.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
